backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.2.4"
description = "image search engine"
authors = ["Qwest1204 <workemailfordaniil@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.112.1"}
pillow = "^10.4.0"
psycopg2 = "^2.9.9"
torch = {version = "^2.4.0+cpu", source = "pytorch_cpu"}
torchvision = {version = "^0.19.0+cpu", source = "pytorch_cpu"}
open-clip-torch = "^2.26.1"
opencv-python = "^4.10.0.84"

[tool.poetry.group.dev.dependencies]
chromadb = "^0.5.5"


[[tool.poetry.source]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
